Optimization for static map done.

+++ includes/game_of_life.h
@@ -6,7 +6,7 @@
 /*   By: dmalesev <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/22 10:26:30 by dmalesev          #+#    #+#             */
-/*   Updated: 2022/06/23 11:26:57 by dmalesev         ###   ########.fr       */
+/*   Updated: 2022/06/23 14:26:24 by dmalesev         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,6 +18,6 @@
 #include <stdlib.h>
 
 void	print_map(uint8_t **map, size_t line_len, size_t lines);
-void	iterate_map(uint8_t **map, size_t line_len, size_t lines, int iters);
+void	iterate_map(uint8_t **map, size_t line_len, size_t lines, long iters);
 
 #endif
